# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth": {
            "post": {
                "description": "Authenticate to get JWT token",
                "produces": [
                    "application/json"
                ],
                "summary": "Authentication",
                "parameters": [
                    {
                        "description": "Username and Password",
                        "name": "UsernamePassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mtauthserver.UsernamePassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token response",
                        "schema": {
                            "$ref": "#/definitions/mtauthserver.ResponseToken"
                        }
                    },
                    "400": {
                        "description": "error response",
                        "schema": {
                            "$ref": "#/definitions/mtauthserver.ResponseAuthError"
                        }
                    },
                    "401": {
                        "description": "error response",
                        "schema": {
                            "$ref": "#/definitions/mtauthserver.ResponseAuthError"
                        }
                    }
                }
            }
        },
        "/introspect": {
            "post": {
                "description": "Check JWT token",
                "produces": [
                    "application/json"
                ],
                "summary": "Introspect",
                "parameters": [
                    {
                        "description": "Token",
                        "name": "Token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mtauthserver.Token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "verify response",
                        "schema": {
                            "$ref": "#/definitions/mtauthserver.ResponseVerify"
                        }
                    },
                    "400": {
                        "description": "error response",
                        "schema": {
                            "$ref": "#/definitions/mtauthserver.ResponseAuthError"
                        }
                    },
                    "401": {
                        "description": "error response",
                        "schema": {
                            "$ref": "#/definitions/mtauthserver.ResponseAuthError"
                        }
                    }
                }
            }
        },
        "/verify/{app_name}": {
            "get": {
                "description": "Verify JWT token",
                "produces": [
                    "application/json"
                ],
                "summary": "Verification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "application name",
                        "name": "app_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "authentication token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "verify response",
                        "schema": {
                            "$ref": "#/definitions/mtauthserver.ResponseVerify"
                        }
                    },
                    "400": {
                        "description": "error response",
                        "schema": {
                            "$ref": "#/definitions/mtauthserver.ResponseError"
                        }
                    },
                    "401": {
                        "description": "error response",
                        "schema": {
                            "$ref": "#/definitions/mtauthserver.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "mtauthserver.ResponseAuthError": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "mtauthserver.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "mtauthserver.ResponseToken": {
            "type": "object",
            "properties": {
                "exp": {
                    "type": "integer"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "mtauthserver.ResponseVerify": {
            "type": "object",
            "properties": {
                "app_name": {
                    "type": "string"
                },
                "exp": {
                    "type": "integer"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "mtauthserver.Token": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "mtauthserver.UsernamePassword": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}